Unified backend/frontend types, schemas, cookies
Frontend and backend now share identical type files.
Mongoose schemas have been updated to stricter specifications.
Build file is now served as app.html instead of index.html, so that
express does not skip the app.get('/', ...) route.
Cookie is not set on '/' path so the app can remember if the user is
logged in.