Message model uses ref now, profile update logic
Converted Message model to use a reference to the User/Guest model.
Now, changes in a user's profile in the database will be reflected in
later messageHistory requests, rather than showing their profile at
that moment in time.
req.user usage has been significantly reduced due to the inability to
update req.user on updateProfile event. Instead, getters on usersOnline
will be used to pull the latest state of the userProfile, since the
alternative is querying the database on every relevant event.